+++ HISTORY.rst
@@ -8,7 +8,7 @@ Changelog
 
 - Implemented real-time collaborative editing via together.js
 - Added pagination in *what changed* page
-- Implemented page move
+- Implemented views to move and delete pages
 
 
 0.3.3 (2014-11-24)